Transaction submission  + tests (#34)
* transaction manager

* Transaction Manager Tests

* pending transactions DAO + Scaffold tests

* PendingTransactionsDao + tests

* Persistent Transaction Manager

+++ ZcashLightClientKit/Repository/PendingTransactionRepository.swift
@@ -8,10 +8,10 @@
 import Foundation
 
 protocol PendingTransactionRepository {
-    func create(_ transaction: PendingTransactionEntity) throws -> Int64
+    func create(_ transaction: PendingTransactionEntity) throws -> Int
     func update(_ transaction: PendingTransactionEntity) throws
     func delete(_ transaction: PendingTransactionEntity) throws
     func cancel(_ transaction: PendingTransactionEntity) throws
-    func find(by id: Int64) throws -> PendingTransactionEntity?
+    func find(by id: Int) throws -> PendingTransactionEntity?
     func getAll() throws -> [PendingTransactionEntity]
 }
